+++ MiniEngine/MiniEngine/vertex.h
@@ -7,7 +7,7 @@ class Vertex
 {
 public:
 	Vertex() : m_translation(Vector3()) {};
-	Vertex(float x, float y, float z) : m_translation(Vector3(x, y, z)) {};
+	Vertex(double x, double y, double z) : m_translation(Vector3(x, y, z)) {};
 	Vertex(Vector3 vec) : m_translation(vec) {};
 
 	Vector3 getTranslation() { return m_translation; }
